+++ src/app/core/footer/footer.component.spec.ts
@@ -1,4 +1,4 @@
-import { cit, create } from '../../../test-support';
+import { cit, create } from '../../../testing';
 import { FooterComponent } from './footer.component';
 
 describe('FooterComponent', () => {
