+++ src/app/shared/model/uploadable.model.spec.ts
@@ -1,5 +1,5 @@
 import { Observable, ConnectableObservable, Subscriber } from 'rxjs';
-import { cms } from '../../../test-support';
+import { cms } from '../../../testing';
 import { Upload } from '../../core/upload/upload.service';
 import { UploadableModel } from './uploadable.model';
