+++ src/app/shared/upload/directives/audio-file.component.spec.ts
@@ -1,4 +1,4 @@
-import { cit, create, stubPipe, By } from '../../../../test-support';
+import { cit, create, stubPipe, By } from '../../../../testing';
 import { AudioFileComponent } from './audio-file.component';
 
 describe('AudioFileComponent', () => {
